Fix model selection bug
Fixed bug where selecting a model in the plot would display the wrong
model if the model index was greater than 200 (the model index would be
x modulu 200). I also made sure that selecting a model index manually
would not override selections done in the plot latter on, by only
changing the session state if the selectbox changed (via a callback).