Fix index bug
I accidentally left over the index argument which I thought could be set
to change the value of the selected option in selectbox. It only changes
the default option, so is useless anyway and also the value we set is
not always what we intend, so it should just be removed.